requirements.txt
gradio
discord.py
openai
cryptography
requests
chromadb
python-chess
english-words
